refactor save_forecast.py into save/save_database.py, save/save_site_database.py, and save/save_csv.py
