+++ solar_consumer/save/__init__.py
@@ -1,0 +1,1 @@
+# Save module for solar consumer
