+++ solar_consumer/save/save_site_database.py
@@ -1,12 +1,10 @@
 from loguru import logger
-from nowcasting_datamodel.save.save import save
 from pvsite_datamodel.write.generation import insert_generation_values
 from pvsite_datamodel.write.forecast import insert_forecast_values
 from pvsite_datamodel.read.site import get_site_by_client_site_name
 from pvsite_datamodel.write.user_and_site import create_site
 from pvsite_datamodel.pydantic_models import PVSiteEditMetadata as PVSite
 from sqlalchemy.orm.session import Session
-import os
 import pandas as pd
 from typing import Optional
 
@@ -252,64 +250,3 @@ def save_forecasts_to_site_db(
         ml_model_version=model_version,
         session=session,
     )
-
-
-def save_forecasts_to_db(forecasts: list, session: Session):
-    """Save forecasts to the database.
-
-    Parameters:
-        forecasts (list): List of forecast objects to save.
-        session (Session): SQLAlchemy session for database access.
-
-    Return:
-        None
-    """
-    # Check if forecasts is empty
-    if not forecasts:
-        logger.warning("No forecasts provided to save!")
-        return
-
-    try:
-        logger.info("Saving forecasts to the database.")
-        save(
-            forecasts=forecasts,
-            session=session,
-        )
-        logger.info(f"Successfully saved {len(forecasts)} forecasts to the database.")
-    except Exception as e:
-        logger.error(f"An error occurred while saving forecasts: {e}")
-        raise e
-
-
-def save_forecasts_to_csv(forecasts: pd.DataFrame, csv_dir: str):
-    """Save forecasts to a CSV file.
-
-    Parameters:
-        forecasts (pd.DataFrame): DataFrame containing forecast data to save.
-        csv_dir (str): Directory to save CSV files.
-
-    Return:
-        None
-    """
-    # Check if forecasts is empty
-    if forecasts.empty:
-        logger.warning("No forecasts provided to save!")
-        return
-
-    if not csv_dir:  # check if directory csv directory provided
-        raise ValueError("CSV directory is not provided for CSV saving.")
-
-    os.makedirs(csv_dir, exist_ok=True)
-    csv_path = os.path.join(csv_dir, "forecast_data.csv")
-
-    try:
-        forecasts.drop(
-            columns=["_sa_instance_state"], errors="ignore", inplace=True
-        )  # Remove SQLAlchemy metadata
-
-        logger.info(f"Saving forecasts to CSV at {csv_path}")
-        forecasts.to_csv(csv_path, index=False)
-        logger.info(f"Successfully saved {len(forecasts)} forecasts to CSV.")
-    except Exception as e:
-        logger.error(f"An error occurred while saving forecasts to CSV: {e}")
-        raise e
