+++ tests/test_real_forecasts.py
@@ -21,10 +21,8 @@ import pandas as pd
 from nowcasting_datamodel.models import ForecastSQL
 from solar_consumer.fetch_data import fetch_data
 from solar_consumer.format_forecast import format_to_forecast_sql
-from solar_consumer.save_forecast import (
-    save_forecasts_to_csv,
-    save_forecasts_to_db,
-)
+from solar_consumer.save.save_csv import save_forecasts_to_csv
+from solar_consumer.save.save_database import save_forecasts_to_db
 
 
 @pytest.mark.integration
